Kleinverbruik webservice
Add a container and automatic deploy for a Stedin kleinverbruik
webservice.

I'll probably rework this container in the future to handle both Stedin
and Liander kleinverbruik because the logic is the same except for small
differences in the input CSV.

Enexis has a webservice so I'm using that to keep Zenmo Zero as
stateless as possible. But it might be handy to integrate it all into
one service which covers all of the Netherlands.
