+++ kleinverbruik/.gitignore
@@ -1,2 +1,3 @@
 node_modules
 stedin_kleinverbruikgegevens_2023-01-01.csv
+stedin_kleinverbruikgegevens_2023-01-01.json
